requirements.txt
Django==4.2.11
djangorestframework==3.15.1
django-filter=24.2
backports.zoneinfo==0.2.1
sqlparse==0.5.0
typing_extensions==4.11.0
asgiref==3.8.1
